refactor: Separate tests
